+++ weblate/addons/models.py
@@ -113,6 +113,7 @@ class AddonsConf(AppConf):
         'weblate.addons.gettext.MsgmergeAddon',
         'weblate.addons.gettext.GettextCustomizeAddon',
         'weblate.addons.cleanup.CleanupAddon',
+        'weblate.addons.discovery.DiscoveryAddon',
         'weblate.addons.flags.SourceEditAddon',
         'weblate.addons.flags.TargetEditAddon',
         'weblate.addons.generate.GenerateFileAddon',