+++ docs/changes.rst
@@ -2,9 +2,11 @@ Changes
 =======
 
 weblate 3.0
-------------
+-----------
 
 Released on ? 2018.
+
+* New addon for automatic component discovery.
 
 weblate 2.20
 ------------